Avoid NPE when there are both lombok tasks and other non-Java compile tasks
When no Java compilation task is matched to a delombok task, do not throw a null pointer exception, since there are legitimate reasons to have non-Java compile tasks.

I also modified the check for what is a delombok task, because the delombok tasks of subprojects are prefixed with the subproject name. The previous version of the plugin also did not support Lombok in subprojects.